ppbackend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
better-profanity==0.7.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==65.0.2
django-filter==24.3
django-jet-reboot==1.3.9
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
idna==3.10
Markdown==3.7
nanoid==2.0.0
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3